tasklist on separated component
