+++ tasklist/package.json
@@ -9,13 +9,15 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "^1.18.1",
+    "@react-native-async-storage/async-storage": "~1.17.3",
     "expo": "^46.0.21",
     "expo-status-bar": "~1.4.4",
-    "expo-updates": "^0.4.2",
-    "react": "18.2.0",
-    "react-dom": "^18.2.0",
-    "react-native": "0.71.3",
+    "expo-updates": "~0.14.7",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-native": "0.69.9",
+    "react-native-scrollable-tab-view": "^1.0.0",
+    "react-native-tab-view": "^3.5.2",
     "react-native-web": "~0.18.11"
   },
   "devDependencies": {
